% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finanzas-dataset.R
\docType{data}
\name{finanzas}
\alias{finanzas}
\title{Registro de egresos/ingresos}
\format{
Un data.frame con 3 variables y observaciones variables

\itemize{
\item{\code{fecha_pago}: Fecha en que el concepto se hace efectivo.}
\item{\code{concepto}: Descriptor del concepto}
\item{\code{valor}: Monto del concepto}
}
}
\usage{
finanzas
}
\description{
Registro de egresos e ingresos con sus respectivas fechas asociadas.
}
\keyword{datasets}
