% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices-package.R
\docType{package}
\name{prices-package}
\alias{prices}
\alias{prices-package}
\title{prices: Calcula Precios para trabajo Tiempo-Dependiente}
\description{
Calcule los precios de los trabajos en función del tiempo para independientes o cualquier otra forma de trabajo que requiera cobrar por hora/sesión.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matcasti/prices}
  \item \url{https://matcasti.github.io/prices/}
  \item Report bugs at \url{https://github.com/matcasti/prices/issues}
}

}
\author{
\strong{Maintainer}: Matías Castillo Aguilar \email{matcasti@umag.cl}

}
\keyword{internal}
