% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices_balance.R
\name{prices_balance}
\alias{prices_balance}
\title{Calcula balances financieros ordenados por fecha}
\usage{
prices_balance(data, fecha, valor, balance_inicial = 0, alcanza = TRUE)
}
\arguments{
\item{data}{Datos. Deben contener como mínimo la fecha (en el cual se asigna el concepto) o número
(que sirva para ordenar la secuencia del balance), así como el valor o monto asociado al egreso/ingreso.
Adicionalmente puede tener una columna con los conceptos (un descriptor), el cual no se verá modificado.}

\item{fecha}{Nombre de la columna con las fechas o número que sirva para ordenar los valores/montos.}

\item{valor}{Nombre de la columna con los valores. Los valores/montos deben ser mayor a cero (> 0)
si son ingresos, y menores a cero (< 0) si son egresos/gastos.}

\item{balance_inicial}{Número indicando el balance inicial al cual se le aplican los conceptos.}

\item{alcanza}{Lógico indicando si el se desea una columna indicando si el balance se mantiene positivo.}
}
\description{
Calcula el balance secuencial ordenado por fecha. Ideal para estimar el balance de un conjunto
de egresos/ingresos.
}
\examples{
prices_balance(finanzas, fecha_pago, valor)

}
