% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices_matrix.R
\name{prices_matrix}
\alias{prices_matrix}
\title{Generador de matrices de precios}
\usage{
prices_matrix(valor_hora, sesiones)
}
\arguments{
\item{valor_hora}{Numérico. Vector con los valores por hora.}

\item{sesiones}{Numérico. Vector con número de sesiones.}
}
\value{
Matriz si alguno de los vectores es de longitud > 1, de lo contrario
devuelve un vector de longitud 1 equivalente a el producto de \code{valor_hora} y \code{sesiones}
}
\description{
Genera una matriz de valores para una combinación de precios por hora
y una cantidad de sesiones en formato vectores, pudiendo ser (idealmente),
de longitud > 1 (devolviendo así una matriz).
}
\examples{

  prices_matrix(
    valor_hora = 10:15,
    sesiones = c(8, 12, 24, 36, 48, 60)
   )

}
