% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices_ranges.R
\name{prices_ranges}
\alias{prices_ranges}
\title{Generador de precios en secuencia}
\usage{
prices_ranges(valor_hora, sesiones = c(4, 8, 12, 24, 36, 48, 60), k = 3)
}
\arguments{
\item{valor_hora}{Númerico de largo 2. Vector con los valores a interpolar.}

\item{sesiones}{Numérico. Vector con número de sesiones.}

\item{k}{Numérico. Número de valores decimales (por defecto 3).}
}
\value{
data.frame con tres columnas: \code{sesiones}, indicando el número
de sesiones, \code{valor_hora}, indicando el valor de la sesión y \code{valor_plan},
indicando el valor final del plan.
}
\description{
Dado un precio mínimo y máximo (en un vector de longitud 2), genera
una secuencia de precios tomando los intervalos regulares desde el
mínimo hasta el máximo dividido en el número total de sesiones dadas.
}
\examples{

  list(
    online_individual     = prices_ranges(valor_hora = c(15, 10)),
    presencial_individual = prices_ranges(valor_hora = c(20, 15)),
    presencial_duo        = prices_ranges(valor_hora = c(30, 20)),
    presencial_trio       = prices_ranges(valor_hora = c(40, 30)),
    presencial_domicilio  = prices_ranges(valor_hora = c(25, 17.5))
  )

}
